Basic working upload of images to cloudinary
TODO
Need to simplify code to reduce the repeating, add the 3rd call to
backend to write the database document, and also switch to
using redux toolkit query.

Also in our first call to our backend, send the form data (product
title and product description), and validate this data on both backend
requests. This way we can return the validation error on the 1st
backend request without uploading anything to cloudinary