Added React Dropzone for Basic Draggable Area Btn
I think though, that later I will want the ability to upload multiple
images to a single product. I need to change the Database to support
this, but also our current UI for adding a new Product in our
Modal/Dialog would not work well for this. We'd need to be able to show
multiple images and allow a user to scroll through them. For this,
I think its best to just use an entire new page instead of a modal.
Another alternative is to make the modal a lot wider so that one side
of it can be allocated to showing drop area + all images + currently
selected image to preview.
I can maybe start with a wider modal, then decide whether I need to
instead turn 'New Product' into a new page

TODO
-Change 'New Product', keeping in mind support for showing drop area +
all images + currently selected image to preview. Try either using a
wider modal, or a brand new page (no modal) for 'New Product'
-better styles on everything
-Figure out how to add/show a form for editing a product: Will I just
inline the JSX, or re-use my custom <Input> somehow, or pop up a modal
everytime I click on Edit?