Fixed some styles in Products. Fetch fixed imgs
Fetch calls fixed to account for changes in DB for multiple images.
The actual implemetnation of fulkly supporting multiple images still
needs to be worked on (UI and code to send do all cloudinary uploads).
The frontend code to do this though is just a loop. We can also reuse
the 1 presignedURL signature, since it takes 1 hour for the signature
to expire. Example can be found at the bottom of this page:
https://cloudinary.com/documentation/authentication_signatures

TODO
refer to prev commit