Added query error toast on RTK middleware
-In RTK middleware, I also switched to using utility functions from
RTK Query to identify if a query is pending or fufilled or rejected
-Also added barebones profile page, as well as the navbar dropdown menu
when clicking on profile image in navbar now includes user level and
username
-Also Changed New Product submit handler flow to use
await myMutationTrigger.unwrap() instead of not calling unwrap. This
makes it really easy to simply wrap the entire thing in a try/catch,
and any fetch errors will throw (exit to the catch block).
Otherwise if we don't unwrap, we would have to check if(error)return
after every call.

TODO - Here are some todos from before, still TODO
-Fetch calls fixed to account for changes in DB for multiple images.
The actual implemetnation of fullly supporting multiple images still
needs to be worked on (UI and code to send do all cloudinary uploads).
The frontend code to do this though is just a loop. We can also reuse
the 1 presignedURL signature, since it takes 1 hour for the signature
to expire. Example can be found at the bottom of this page:
https://cloudinary.com/documentation/authentication_signatures

-the new product modal/img shows the image in the same display form as
Products UI does, however the css/tailwind is different. I wrestled with
it in New Products, but I think it ended up much simpler than in
Products. Review if it really was better, and see if I cann change
Products tailwind to this approach of css/tailwind