Added Category List & Management
TODO
-refer to prev commits
-Maybe make it so categories created by owner can only be managed by
a user with owner role. Maybe also makeit possible to let owners assign
an 'owner-only' property to a category created by admins, or something
like this
- In the backend for categories, when a catgory is deleted, then
query all products in the database and set their Category to undefined.
Also look at what the default behavior of the DB would be if we dont
do this