Adding Categories Input to Products - in progress
UI looks good/ok and works. Now need to include the categories selected
as part of the form payload we currently send. We can do this manually,
but I think its best to either create a hidden input that we populate or
use a react-hook-form custom input. The objective is ease of sending
info as part of the form + ALSO accessibility

ALSO switched to using vite-plugin-svgr for SVGs. This lets us easily
import svgs as react components, so that we can change the svg's colors
from the JSX. Some SVGs might need their contents inside the file
changed to 'fill=currentColor'

TODO
-Add categories input array as part of the form data, as described above
(prob use react-hook-form custom input)
-few cleanup comments etc
-refer to many prev commits