Added ProductModalEdit except for requests
I finalized ProductModalNew. Everything accounts for the potential to
add/remove multiple images. Just need to change the UI to show these
images. Right now the 'maxFiles' is set to 1, to restrict to 1 image max

TODO
-test ProductModalNew, see if everything frontend + backend still works.
-for ProductModalEdit, use as reference, the request code we wrote for
ProductModalNew. I just now realized though, that the code we wrote for
this in ProductModalNew, was assuming that we might be editing stuff(?),
AKA, we assumed that we might have image object with an existing '.url'
property, which CAN ONLY HAPPEN in EditModal, not in NewModal.
Check this, maybe there are more spots where we were assuming that we
were coding as if it was the EditModal that opens up with starting
values
-After all of above, then can start displaying extra Images in the UI.
Aka increase the maxFiles to 4 or something, so that they are shown in
the UI and can be removed, and then test all modals again frontend +
backend
-refer to rpev commmits