Finalized ProductsModal both New and Edit
TODO
-Change the fileLimit on ProductModal from 1 to prob 4. Change the UI
in the modals and in the ProductDetail, so that they support showing
all the images and the Edit Modal supports changing their order.
All the code for this in the backend and frontend has already been
written to account for multiple images, the only things left are to
test that code + add the UI functionalities described above.
-refer to prev commits