Moved image state into form state. Fixed avatar
Image and Category states now is part of the react-hook-form initial
state, instead of each having its own useState
Google Avatar sometimes not loading was fixed by adding html attribute:
referrerPolicy='no-referrer' to the img element. Apparently this google
pfp not loading problem only happens when hosting locally (intended by
google), but I didnt test if it happend when deployed or not. Either
way, this seems to fix it