Added multiple images UI in all places V1
TODO
-maybe add hover to select in productModal, not sure tho but can try it,
its simply changing the onClick to OnMouseEnter
-probably change selection color in both ProductModal and ProductDetail
-see if CategoryModalEdit should replicate the reset logic that
we now use in ProductModalEdit
-right now we use our 100% or 50% of out image sort logic in some
places, I think these are all necessary, though not 100% sure:
1. in our schema so that the backend always makes sure any input array
of images (even with bad input), will be numbered 1 through X
2. in the backend productEdit, because we seprate the image array into 3
arrays and then combine it at the end, so we need to sort by order
3. When removing an image in the ProductModal, because we need to make
sure our images are numbered 1 through X in sequence, and not sorting
them would leave gaps
4. in /shop, which makes sense to me bc we cant rely
on the array always coming in sorted, and should instead sort
manually  based on the order property(?). I guess we can maybe force
our backend to always sort the product images when retrieving, but I
think we shouldnt. We would have to ensure this everywhere, and I
think just relying 100% on .order makes a lot more sense